logging refactor
