+++ walax/src/model/DjangoQuery.js
@@ -1,6 +1,6 @@
 import w from '../Walax'
 
-import Logger from './Logger'
+import Logger from '../control/Logger'
 const { d, a, e, i } = Logger.daei('Auth')
 
 class DjangoQueryProxy {
